Initial commit.
The current code solves the 1D Schroedinger equation for a potential
V(x) to be given as an expansion in Hermite polynomials.  The solution
is written as a linear combination of eigenfunctions of the harmonic
oscillator (the frequency of which needs to be provided); the
expansion order is adjusted to converge the variational energy.

TO DO:
* Read V(x) as a natural polynomial.
* Automatically pick "best" omega (need criterion).
* Find why LAPACK eigenvalue solvers hang for matrices of size
  202x202.
* Implement solving for the quadrature grid.
* Implement testing the quadrature grid.