Centre shifting; robust determination of ucentre and omega.
Three methods to determine ucentre and omega are implented:
* ucentre = <u> self-consistently, omega minimizes virial ratio error
* ucentre = <u> self-consistently, omega minimizes variational energy
* ucentre = <u> and omega = 1/(2<u^2>) self-consistently
Internal parameters SET_OMEGA_BY_X2 and MINIMIZE_E control which one
is used.  The third method above tends to give lower omega - while
this implies the need for higher expansions orders in solving the
Schroedinger equation, it also improves the numerics of the grid
solver, so this is the default.