Added two plot functions; bumped MAX_NGRID to 12.
New plot functions are polynomials, which are particularly relevant
(these model band gaps).  These simply serve to verify that
polynomials are integrated exactly by a grid with a matching number
of points.

A larger MAX_NGRID allows visualizing the overall convergence pattern
of the Lorentzian, which is non-monotonic -- truncating the plot at
order 8 gives the impression that the grid integral converges to
the wrong value.